+++ src/main/java/dev/willyelton/crystal_tools/common/events/ShieldBlockEvent.java
@@ -2,14 +2,29 @@ package dev.willyelton.crystal_tools.common.events;
 
 import dev.willyelton.crystal_tools.CrystalTools;
 import dev.willyelton.crystal_tools.common.components.DataComponents;
+import dev.willyelton.crystal_tools.common.components.EffectData;
 import dev.willyelton.crystal_tools.common.levelable.tool.CrystalShield;
+import net.minecraft.core.Holder;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.network.protocol.game.ClientboundSoundPacket;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.projectile.Projectile;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.event.entity.living.LivingShieldBlockEvent;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
 
 @EventBusSubscriber(modid = CrystalTools.MODID)
 public class ShieldBlockEvent {
@@ -26,29 +41,46 @@ public class ShieldBlockEvent {
                 crystalShield.addExp(blockingEntity.getUseItem(), blockingEntity.level(), blockingEntity.getOnPos(), blockingEntity, (int) Math.ceil(event.getBlockedDamage()));
 
                 if (attackingEntity instanceof LivingEntity targetEntity) {
-//                    handleTargetEffects(targetEntity, crystalShield, blockingEntity);
+                    handleTargetEffects(targetEntity, stack, blockingEntity, blockingEntity.level());
                 }
 
                 if (directEntity instanceof Projectile projectile) {
                     int target = stack.getOrDefault(DataComponents.ENTITY_TARGET, -1);
                     if (target != -1) {
-                        ServerTickEvent.startTracking(attackingEntity.level(), projectile.getId(), target);
+                        LevelTickEvent.startTracking(attackingEntity.level(), projectile.getId(), target, projectile.getDeltaMovement().length());
                     }
                 }
             }
         }
     }
 
-    private static void handleTargetEffects(LivingEntity target, CrystalShield shieldItem, LivingEntity player) {
+    private static void handleTargetEffects(LivingEntity target, ItemStack stack, LivingEntity blocker, Level level) {
         // Fire
-        target.igniteForTicks(100);
+        if (stack.getOrDefault(DataComponents.FLAMING_SHIELD, false)) {
+            target.igniteForTicks(100);
+        }
 
-        // Push back (add a little vertical here)
-        target.push(target.getPosition(0).subtract(player.getPosition(0)).add(0, 0.5, 0));
+        if (stack.getOrDefault(DataComponents.SHIELD_KNOCKBACK, false)) {
+            if (blocker instanceof ServerPlayer player) {
+                player.connection.send(new ClientboundSoundPacket(Holder.direct(SoundEvents.WIND_CHARGE_BURST.value()), SoundSource.PLAYERS, blocker.getX(), blocker.getY(), blocker.getZ(), 2F, (level.random.nextFloat() - level.random.nextFloat()) * 1.4F + 2.0F, level.getRandom().nextLong()));
+            }
+            target.push(target.getPosition(0).subtract(blocker.getPosition(0)).add(0, 0.5, 0));
+        }
+
+        int thorns = stack.getOrDefault(DataComponents.SHIELD_THORNS, 0);
+
+        if (thorns > 0 && Math.random() < thorns * 0.25) {
+            target.hurt(level.damageSources().thorns(blocker), 3);
+        }
 
         // Mob Effects
-
-        // Thorns
-        // Vanilla is level * .15 chance to do 1-4 damage
+        List<EffectData> effects = stack.getOrDefault(DataComponents.EFFECTS, Collections.emptyList());
+        for (EffectData effect : effects) {
+            Optional<Holder.Reference<MobEffect>> mobEffect = BuiltInRegistries.MOB_EFFECT.getHolder(ResourceLocation.withDefaultNamespace(effect.resourceLocation()));
+            if (mobEffect.isPresent()) {
+                MobEffectInstance instance = new MobEffectInstance(mobEffect.get(), 100, 1, false, true);
+                target.addEffect(instance);
+            }
+        }
     }
 }