+++ src/main/java/dev/willyelton/crystal_tools/common/levelable/tool/CrystalShield.java
@@ -1,20 +1,35 @@
 package dev.willyelton.crystal_tools.common.levelable.tool;
 
+import dev.willyelton.crystal_tools.Registration;
+import dev.willyelton.crystal_tools.common.components.DataComponents;
+import dev.willyelton.crystal_tools.common.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.common.levelable.EntityTargeter;
 import dev.willyelton.crystal_tools.common.levelable.LevelableItem;
+import dev.willyelton.crystal_tools.utils.ToolUtils;
 import net.minecraft.network.chat.Component;
+import net.minecraft.util.Mth;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EquipmentSlotGroup;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ShieldItem;
 import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.component.ItemAttributeModifiers;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BannerPatternLayers;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
+import java.util.function.Consumer;
+
+import static dev.willyelton.crystal_tools.common.levelable.armor.LevelableArmor.ARMOR_ID;
+import static dev.willyelton.crystal_tools.common.levelable.tool.LevelableTool.ATTACK_DAMAGE_ID;
 
 public class CrystalShield extends ShieldItem implements LevelableItem, EntityTargeter {
 
@@ -32,9 +47,7 @@ public class CrystalShield extends ShieldItem implements LevelableItem, EntityTa
 
     @Override
     public void onUseTick(Level level, LivingEntity livingEntity, ItemStack stack, int remainingUseDuration) {
-//        if (level.getGameTime() % 10 == 0) {
-            refreshTarget(stack, level, livingEntity);
-//        }
+        refreshTarget(stack, level, livingEntity);
 
         super.onUseTick(level, livingEntity, stack, remainingUseDuration);
     }
@@ -56,17 +69,92 @@ public class CrystalShield extends ShieldItem implements LevelableItem, EntityTa
     }
 
     @Override
-    public int getMaxDamage(ItemStack stack) {
-        return super.getMaxDamage(stack);
-    }
-
-    @Override
     public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> tooltipComponents, TooltipFlag tooltipFlag) {
         appendLevelableHoverText(stack, tooltipComponents, this, tooltipFlag);
     }
 
     @Override
+    public void addAdditionalTooltips(ItemStack stack, List<Component> components, LevelableItem item) {
+        int totemSlots = stack.getOrDefault(DataComponents.TOTEM_SLOTS, 0);
+        if (totemSlots > 0) {
+            components.add(Component.literal(String.format("%d/%d Totems of Undying", stack.getOrDefault(DataComponents.FILLED_TOTEM_SLOTS, 0), totemSlots)));
+        }
+    }
+
+    @Override
     public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack, boolean slotChanged) {
         return !ItemStack.isSameItem(oldStack, newStack);
+    }
+
+    @Override
+    public int getMaxDamage(ItemStack stack) {
+        int bonusDurability = stack.getOrDefault(DataComponents.DURABILITY_BONUS, 0);
+        return INITIAL_TIER.getUses() + bonusDurability;
+    }
+
+    @Override
+    public <T extends LivingEntity> int damageItem(ItemStack stack, int amount, T entity, Consumer<Item> onBroken) {
+        int durability = this.getMaxDamage(stack) - stack.getDamageValue();
+
+        if (durability - amount <= 0) {
+            return 0;
+        } else {
+            return amount;
+        }
+    }
+
+    @Override
+    public boolean isFoil(ItemStack stack) {
+        return false;
+    }
+
+    @Override
+    public int getBarWidth(ItemStack itemStack) {
+        return Math.round(13.0F - (float) itemStack.getDamageValue() * 13.0F / (float) itemStack.getMaxDamage());
+    }
+
+    @Override
+    public int getBarColor(ItemStack itemStack) {
+        float f = Math.max(0.0F, ((float)itemStack.getMaxDamage() - (float)itemStack.getDamageValue()) / (float) itemStack.getMaxDamage());
+        return Mth.hsvToRgb(f / 3.0F, 1.0F, 1.0F);
+    }
+
+    @Override
+    public void inventoryTick(ItemStack stack, Level level, Entity entity, int slotId, boolean isSelected) {
+        levelableInventoryTick(stack, level, entity, slotId, isSelected, 1);
+    }
+
+    @Override
+    public boolean isValidRepairItem(ItemStack tool, ItemStack repairItem) {
+        return repairItem.is(Registration.CRYSTAL.get());
+    }
+
+    @Override
+    public boolean isEnchantable(@NotNull ItemStack stack) {
+        return CrystalToolsConfig.ENCHANT_TOOLS.get();
+    }
+
+    @Override
+    public boolean isBookEnchantable(ItemStack stack, ItemStack book) {
+        return CrystalToolsConfig.ENCHANT_TOOLS.get();
+    }
+
+    @Override
+    public ItemAttributeModifiers getLevelableAttributeModifiers(ItemStack stack) {
+        ItemAttributeModifiers.Builder builder = ItemAttributeModifiers.builder();
+        if (!ToolUtils.isBroken(stack)) {
+            int armor = stack.getOrDefault(DataComponents.ARMOR_BONUS, 0);
+
+            if (armor > 0) {
+                builder.add(Attributes.ARMOR, new AttributeModifier(ARMOR_ID, armor, AttributeModifier.Operation.ADD_VALUE), EquipmentSlotGroup.OFFHAND);
+            }
+
+            float attackDamage = stack.getOrDefault(DataComponents.DAMAGE_BONUS, 0F);
+            if (attackDamage > 0) {
+                builder.add(Attributes.ATTACK_DAMAGE, new AttributeModifier(ATTACK_DAMAGE_ID, attackDamage, AttributeModifier.Operation.ADD_VALUE), EquipmentSlotGroup.MAINHAND);
+            }
+        }
+
+        return builder.build();
     }
 }